my-react-app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
psycopg2-binary
google-generativeai